Split model page layout sections
